+++ src/index.ts
@@ -710,6 +710,7 @@ export * from './mongo/statistics';
 export * from './oauth/microsoft';
 export * from './queues/index';
 export * from './schemas';
+export * from './services/feedback';
 export * from './statistics';
 export * from './streams';
 export * from './strings';